CXX-1600 Fix 'append' is ambiguous using libc++
The issue CXX-1600 is for different file but the same error.

+++ src/mongocxx/collection.cpp
@@ -230,7 +230,7 @@ bsoncxx::document::value as_bson(const mongocxx::options::change_stream& cs) {
         if ((count < 0) || (count >= std::numeric_limits<std::uint32_t>::max())) {
             throw mongocxx::logic_error{mongocxx::error_code::k_invalid_parameter};
         }
-        out.append(bsoncxx::builder::basic::kvp("maxAwaitTimeMS", count));
+        out.append(bsoncxx::builder::basic::kvp("maxAwaitTimeMS", static_cast<int64_t>(count)));
     }
 
     return out.extract();
